Add files via upload
Sample "Attack to Control" mapping that I've built out before. I'll add this shortly. The idea is to understand what controls you have to actually stop/report on cybersecurity attacks. This is kind of a derivative of the Mitre Attack Framework.

I'll create a guide/walkthrough on how to customize this down the road.